requirements.txt
SQLAlchemy==2.0.23
openai==1.54.4
python-dotenv==1.0.1
python-telegram-bot==21.5